Added quiz results to the database and methods to retrieve them, and created 'models' folder for question.py and database.py
